+++ deerlab/fitmultimodel.py
@@ -102,12 +102,18 @@ def fitmultimodel(V, Kmodel, r, model, maxModels, method='aic', lb=None, ub=None
     Kparam : ndarray
         Fitted kernel parameters.
     
+    V : ndarray or list thereof
+        Fitted dipolar signal(s).
+
     Puncert : :ref:`UQResult`
         Covariance-based uncertainty quantification of the fitted distance distribution
     
     paramUncert : :ref:`UQResult`
         Covariance-based uncertainty quantification of the fitted parameters
-    
+
+    Vuncert : ndarray or list thereof
+        Covariance-based uncertainty quantification of the fitted dipolar signal(s).
+
     Nopt : int scalar
         Optimized number of components in model.
     
@@ -421,12 +427,14 @@ def fitmultimodel(V, Kmodel, r, model, maxModels, method='aic', lb=None, ub=None
         # Separable non-linear least-squares (SNLLS) fit
         scale = 1e2
         fit = dl.snlls(V*scale,Knonlin,par0,nlin_lb,nlin_ub,lin_lb,lin_ub, 
-                        weights=weights, reg=False, uq=False, nonlin_tol=tol, nonlin_maxiter=maxiter)
+                        weights=weights, reg=False, nonlin_tol=tol, nonlin_maxiter=maxiter)
         pnonlin = fit.nonlin
         plin = fit.lin
         par_prev = pnonlin
 
-        plin = plin/scale
+        plin /= scale
+        fit.model /= scale
+        fit.modelUncert = fit.modelUncert.propagate(lambda x: x/scale)
 
         # Store the fitted parameters
         pnonlin_.append(pnonlin)
@@ -451,6 +459,7 @@ def fitmultimodel(V, Kmodel, r, model, maxModels, method='aic', lb=None, ub=None
         lin_ub_.append(lin_ub)
         lin_lb_.append(lin_lb-1) # rescale to zero
         fits.append(fit)
+
     # Select the optimal model
     # ========================
     Peval = Pfit
@@ -523,8 +532,8 @@ def fitmultimodel(V, Kmodel, r, model, maxModels, method='aic', lb=None, ub=None
         fig = _plot(Vsubsets,V,Vfit,r,Pfit,Puq,fcnals,maxModels,method,uq,show)
         return fig
 
-    return FitResult(P=Pfit, Pparam=fitparam_P, Kparam=fitparam_K, amps=fitparam_amp, Puncert=Puq, 
-                    paramUncert=paramuq, selfun=fcnals, Nopt=Nopt, Pn=Peval, scale=scales, plot=plotfcn,
+    return FitResult(P=Pfit, Pparam=fitparam_P, Kparam=fitparam_K, amps=fitparam_amp, V=fit.model, Puncert=Puq, 
+                    paramUncert=paramuq, Vuncert=fit.modelUncert, selfun=fcnals, Nopt=Nopt, Pn=Peval, scale=scales, plot=plotfcn,
                     stats=stats, cost=fit.cost, residuals=fit.residuals, success=fit.success)
 # =========================================================================
 
@@ -540,7 +549,7 @@ def _plot(Vsubsets,V,Vfit,r,Pfit,Puq,fcnals,maxModels,method,uq,show):
         axs[i].plot(Vfit[subset],'tab:blue')
         axs[i].grid(alpha=0.3)
         axs[i].set_xlabel('Array Elements')
-        axs[i].set_ylabel('V[{}]'.format(i))
+        axs[i].set_ylabel(f'V[{i}]')
         axs[i].legend(('Data','Fit'))
 
     if uq:
@@ -567,7 +576,7 @@ def _plot(Vsubsets,V,Vfit,r,Pfit,Puq,fcnals,maxModels,method,uq,show):
     ax = plt.subplot(nSignals+1,2,2*(nSignals+1))
     ax.bar(np.arange(maxModels)+1,np.log10(1 + dfcnals + abs(min(dfcnals))),facecolor='tab:blue',alpha=0.6)
     ax.grid(alpha=0.3)
-    ax.set_ylabel('log10 Δ{}'.format(method.upper()))
+    ax.set_ylabel(f'log10 Δ{method.upper()}')
     ax.set_xlabel('Number of components')
     axs = np.append(axs,ax)